♻️ refactor(dal): imported-game — typed throws (slice 10)
Converts the imported-game vertical and fixes a critical class-identity
bug surfaced during the slice.

- importedGameRepository: P2025 catch translated to NotFoundError with
  operation-specific context. Imports from @/shared/lib/errors (the
  canonical class) — not the legacy data-access-layer/repository/errors,
  which defines a structurally-similar but distinct class.
- ImportedGameService: drops ServiceResult / serviceSuccess /
  handleServiceError. dismissImportedGame returns void;
  findImportedGameById returns Promise<ImportedGame | null>;
  updateImportedGameStatus throws NotFoundError on missing rows.
- imported-games.handler.ts uses mapErrorToHandlerResult.
- Consumers: dismiss action direct-await; import-game-to-library use-case
  has its ImportedGameService side converted (Steam + IGDB still on
  ServiceResult here, addressed in slices 11 / 12).

Class-identity follow-up: journal-repository.ts and
journal-service.unit.test.ts were importing NotFoundError from the legacy
data-access-layer/repository/errors instead of @/shared/lib/errors.
Different prototype chains → instanceof NotFoundError checks in journal
RSC pages' notFound() redirects would have silently failed and surfaced
as generic 500s instead of 404s. Cross-codebase audit confirms no other
stragglers; both files now use the canonical class.
